refactor: recreate presentation projects
