+++ Kysect.AssignmentReporter.sln
@@ -3,10 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.1.31903.286
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kysect.AssignmentReporter.Polygon", "Kysect.AssignmentReporter.Polygon\Kysect.AssignmentReporter.Polygon.csproj", "{F5FD6D08-101A-4566-A3B8-A9602C0F169F}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.AssignmentReporter.Plugin", "Kysect.AssignmentReporter.Plugin\Kysect.AssignmentReporter.Plugin.csproj", "{4646A54C-D469-4BA2-B89F-DFD77CB8129E}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{8B35351E-6CA6-432D-A3C9-ED533D905423}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
@@ -40,6 +36,14 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kysect.AssignmentReporter.I
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.AssignmentReporter.Integration.OfficeIntegration", "Integrations\Kysect.AssignmentReporter.Integration.OfficeIntegration\Kysect.AssignmentReporter.Integration.OfficeIntegration.csproj", "{4715D120-C42A-4B0C-B7F1-380B8B952FEF}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Presentation", "Presentation", "{CC00B921-A789-4F15-96C1-FB212583FC98}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.AssignmentReporter.Plugin", "Presentation\Kysect.AssignmentReporter.Plugin\Kysect.AssignmentReporter.Plugin.csproj", "{4646A54C-D469-4BA2-B89F-DFD77CB8129E}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kysect.AssignmentReporter.Polygon", "Presentation\Kysect.AssignmentReporter.Polygon\Kysect.AssignmentReporter.Polygon.csproj", "{4735029D-D86F-4576-BCBE-BDF5A5CDEB8D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.AssignmentReporter.Api", "Presentation\Kysect.AssignmentReporter.Api\Kysect.AssignmentReporter.Api.csproj", "{72B6AA82-F53F-4815-B8B6-0B6580437668}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		CI|Any CPU = CI|Any CPU
@@ -47,17 +51,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F5FD6D08-101A-4566-A3B8-A9602C0F169F}.CI|Any CPU.ActiveCfg = CI|Any CPU
-		{F5FD6D08-101A-4566-A3B8-A9602C0F169F}.CI|Any CPU.Build.0 = CI|Any CPU
-		{F5FD6D08-101A-4566-A3B8-A9602C0F169F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F5FD6D08-101A-4566-A3B8-A9602C0F169F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F5FD6D08-101A-4566-A3B8-A9602C0F169F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F5FD6D08-101A-4566-A3B8-A9602C0F169F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4646A54C-D469-4BA2-B89F-DFD77CB8129E}.CI|Any CPU.ActiveCfg = CI|Any CPU
-		{4646A54C-D469-4BA2-B89F-DFD77CB8129E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4646A54C-D469-4BA2-B89F-DFD77CB8129E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4646A54C-D469-4BA2-B89F-DFD77CB8129E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4646A54C-D469-4BA2-B89F-DFD77CB8129E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{4B3A9962-23B6-4B1A-8370-51A68373AB6A}.CI|Any CPU.ActiveCfg = CI|Any CPU
 		{4B3A9962-23B6-4B1A-8370-51A68373AB6A}.CI|Any CPU.Build.0 = CI|Any CPU
 		{4B3A9962-23B6-4B1A-8370-51A68373AB6A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
@@ -118,6 +111,24 @@ Global
 		{4715D120-C42A-4B0C-B7F1-380B8B952FEF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4715D120-C42A-4B0C-B7F1-380B8B952FEF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4715D120-C42A-4B0C-B7F1-380B8B952FEF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4646A54C-D469-4BA2-B89F-DFD77CB8129E}.CI|Any CPU.ActiveCfg = CI|Any CPU
+		{4646A54C-D469-4BA2-B89F-DFD77CB8129E}.CI|Any CPU.Build.0 = CI|Any CPU
+		{4646A54C-D469-4BA2-B89F-DFD77CB8129E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4646A54C-D469-4BA2-B89F-DFD77CB8129E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4646A54C-D469-4BA2-B89F-DFD77CB8129E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4646A54C-D469-4BA2-B89F-DFD77CB8129E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4735029D-D86F-4576-BCBE-BDF5A5CDEB8D}.CI|Any CPU.ActiveCfg = CI|Any CPU
+		{4735029D-D86F-4576-BCBE-BDF5A5CDEB8D}.CI|Any CPU.Build.0 = CI|Any CPU
+		{4735029D-D86F-4576-BCBE-BDF5A5CDEB8D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4735029D-D86F-4576-BCBE-BDF5A5CDEB8D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4735029D-D86F-4576-BCBE-BDF5A5CDEB8D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4735029D-D86F-4576-BCBE-BDF5A5CDEB8D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{72B6AA82-F53F-4815-B8B6-0B6580437668}.CI|Any CPU.ActiveCfg = Debug|Any CPU
+		{72B6AA82-F53F-4815-B8B6-0B6580437668}.CI|Any CPU.Build.0 = Debug|Any CPU
+		{72B6AA82-F53F-4815-B8B6-0B6580437668}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{72B6AA82-F53F-4815-B8B6-0B6580437668}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{72B6AA82-F53F-4815-B8B6-0B6580437668}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{72B6AA82-F53F-4815-B8B6-0B6580437668}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -133,6 +144,9 @@ Global
 		{0110DBEF-A9DB-48EF-885B-6412F7DFE70C} = {34D4CDD8-2C59-4EE1-AA87-EC71FCB65206}
 		{FD9B23A6-43E5-4F3D-9C80-7B668ED7701E} = {A2FD4F22-5500-4E13-A47D-5689C6023877}
 		{4715D120-C42A-4B0C-B7F1-380B8B952FEF} = {A2FD4F22-5500-4E13-A47D-5689C6023877}
+		{4646A54C-D469-4BA2-B89F-DFD77CB8129E} = {CC00B921-A789-4F15-96C1-FB212583FC98}
+		{4735029D-D86F-4576-BCBE-BDF5A5CDEB8D} = {CC00B921-A789-4F15-96C1-FB212583FC98}
+		{72B6AA82-F53F-4815-B8B6-0B6580437668} = {CC00B921-A789-4F15-96C1-FB212583FC98}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {91149FF2-74CE-429E-9EE9-F5DAED9756B0}
