+++ Presentation/Kysect.AssignmentReporter.Api/Kysect.AssignmentReporter.Api.csproj
@@ -1,0 +1,17 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <FrameworkReference Include="Microsoft.AspNetCore.App" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\Application\Kysect.AssignmentReporter.Application\Kysect.AssignmentReporter.Application.csproj" />
+  </ItemGroup>
+
+</Project>
