+++ Presentation/Kysect.AssignmentReporter.Plugin/AssignmentReporter.snk
