+++ Presentation/Kysect.AssignmentReporter.Plugin/Kysect.AssignmentReporter.Plugin.csproj
@@ -133,16 +133,6 @@
     </Page>
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Kysect.AssignmentReporter.OfficeIntegration\Kysect.AssignmentReporter.OfficeIntegration.csproj">
-      <Project>{4715d120-c42a-4b0c-b7f1-380b8b952fef}</Project>
-      <Name>Kysect.AssignmentReporter.OfficeIntegration</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\Kysect.AssignmentReporter\Kysect.AssignmentReporter.ReportGenerator.csproj">
-      <Project>{163e8146-c1f5-4959-9df1-7ef4fec2f0ad}</Project>
-      <Name>Kysect.AssignmentReporter.ReportGenerator</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
     <Content Include="Resources\ToolWindowCommand.png" />
     <Content Include="Resources\ToolWindowPackage.ico" />
     <VSCTCompile Include="ToolWindowPackage.vsct">
@@ -155,7 +145,16 @@
       <LastGenOutput>Resource.Designer.cs</LastGenOutput>
     </EmbeddedResource>
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <ProjectReference Include="..\..\Domain\Kysect.AssignmentReporter.ReportGenerator\Kysect.AssignmentReporter.ReportGenerator.csproj">
+      <Project>{478c5f49-99fa-4d09-b32e-d28ea07a4548}</Project>
+      <Name>Kysect.AssignmentReporter.ReportGenerator</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\..\Integrations\Kysect.AssignmentReporter.Integration.OfficeIntegration\Kysect.AssignmentReporter.Integration.OfficeIntegration.csproj">
+      <Project>{4715d120-c42a-4b0c-b7f1-380b8b952fef}</Project>
+      <Name>Kysect.AssignmentReporter.Integration.OfficeIntegration</Name>
+    </ProjectReference>
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(VSToolsPath)\VSSDK\Microsoft.VsSDK.targets" Condition="'$(VSToolsPath)' != ''" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
