+++ Presentation/Kysect.AssignmentReporter.Plugin/Kysect.AssignmentReporter.PluginPackage.cs
