+++ Presentation/Kysect.AssignmentReporter.Plugin/Properties/AssemblyInfo.cs
