+++ Presentation/Kysect.AssignmentReporter.Plugin/Resource.resx
