+++ Presentation/Kysect.AssignmentReporter.Plugin/Resources/ToolWindowCommand.png
