+++ Presentation/Kysect.AssignmentReporter.Plugin/Resources/ToolWindowPackage.ico
