+++ Presentation/Kysect.AssignmentReporter.Plugin/ToolWindowCommand.cs
