+++ Presentation/Kysect.AssignmentReporter.Plugin/ToolWindowPackage.vsct
