+++ Presentation/Kysect.AssignmentReporter.Plugin/ViewModel/CoverPageSettingsViewModel.cs
