+++ Presentation/Kysect.AssignmentReporter.Plugin/ViewModel/MvvmBase/BaseViewModel.cs
