+++ Presentation/Kysect.AssignmentReporter.Plugin/ViewModel/MvvmBase/RelayCommand.cs
