+++ Presentation/Kysect.AssignmentReporter.Plugin/ViewModel/SearchSettingsViewModel.cs
