+++ Presentation/Kysect.AssignmentReporter.Plugin/Windows/CoverPageSettingsWindow.xaml.cs
