+++ Presentation/Kysect.AssignmentReporter.Plugin/Windows/SearchSettingsWindow.xaml
