+++ Presentation/Kysect.AssignmentReporter.Plugin/Windows/ToolWindowControl.xaml.cs
