+++ Presentation/Kysect.AssignmentReporter.Plugin/source.extension.vsixmanifest
