+++ Presentation/Kysect.AssignmentReporter.Polygon/Credentials.cs
