+++ Presentation/Kysect.AssignmentReporter.Polygon/Program.cs
